improve push
Add a filter system on the client side to accept or reject entries from the
push.
Add a better storage system for objects on the server side to limit the number
of object files per directory.
Add a branch support on server side with possibility to create branches with
names and several snapshots per branch, each referencing a root object.