feat(tree): rebase GlobPattern on pcre2
Change the implementation of GlobPattern objects so they use the PCRE2
instead of standard fnmatch() routine.
The GlobPattern convert the extended glob expression into a Perl regex
before to use PCRE2 to compile this regex.

Also add the ability to tranfer Filter and Pattern objects through a
stream with the help of FilterCodec and PatternCodec objects. Because
Pattern and Filter can go through a stream, they are now part of the
tree category (instead of plan category).