feat(io): add an SshChannel built atop PipeChannel
Add an SshChannel class which, just like FileChannel is a building class
for a PipeChannel with a specific command.
SshChannel builds the command by parsing the location to extract the
user, the server address, the port and the remote path.
SshChannel also takes an optional remote command to execute instead of
`synctl server`.