feat(ui): add a path resolution function
Add a function to resolve a path starting with a tilde (~) character.
If the first component is a single tilde, the function resolves it into
the current user's home directory.
If something other than a slash (/) follows the tilde, than it is a name
and the function resolves the first path component into the home
directory of a user with this name.