fix(plan): fix send/pull operations failure on armv7 architectures
Before this fix, an implicit conversion from size_t to uint64_t used to
occur during a send. This is harmless on x86_64 because the two types
are the same but not on armv7. This cast was causing pull tests to fail
on the armv7 architecture.
This commit fix this issue.