feat(tree): add an opcode field in Directory_1 children
Add a field in the Directory_1::Entry structure that get read/written to
indicate the opcode of the pointed children. The goal is to move this
indication from the object file header to the directory content since a
same content can stand for different object types (typically the empty
content).