feat(plan): add an Explorer object
Add the Explorer object to explore entries already present in the
repository and get, given a snapshot root and a path, the reference of
the corresponding entry as well as its opcode and other properties such
as stat and xattributes.
The Explorer can also list the children of an explored entry.