fix(tree): make the stat decoding of Directory_1::Entry optional
Change the implementation of Directory_1::Entry. It is now a proxy for
the internal representation of entries. This internal representation is
portable and does not need for user and group names to be converted in a
proper uid / gid.

This removes the necessity for the server side to make the conversion
between names and ids, which is possibly meaningless in a server with
different users than the clients.