fix(plan): fix communication issues on pull termination
Fix communication errors on the termination of a pull operation. The
server was silently failing on a SIGPIPE because of a useless opcode
sent at the end of Send.
Also, the client could terminate before the server finished to change
the branch if needed.
Now, closing a PipeChannel also means to wait for the child process to
terminate.